% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_calls.R
\name{request_access_token}
\alias{request_access_token}
\title{Request Access Token}
\usage{
request_access_token(
  user_pwd,
  url = "https://api.paypal.com/v1/oauth2/token",
  print_expiry_time = FALSE,
  verbose = FALSE,
  return_token_only = FALSE
)
}
\arguments{
\item{user_pwd}{The user credentials, consisting of a Client ID and Secret.
These should be concatenated as <Cient_ID:Secret>.}

\item{url}{Url to the paypal API, should not neccessarily be changed.}

\item{print_expiry_time}{TRUE prints the time of expiry for the token, default set to FALSE.}

\item{verbose}{TRUE prints response statements from the API, default set to FALSE.}

\item{return_token_only}{TRUE only returns the token, else a list with more attributes is returned,
default is set to FALSE.}
}
\value{
The access token only or a list with the access token and more attributes.
}
\description{
This function uses the credentials to request an access token
needed later in the transaction call. The token expires at some point.
}
