% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing_functions.R
\name{transaction_list_to_df}
\alias{transaction_list_to_df}
\title{Converts Transaction List to Dataframe}
\usage{
transaction_list_to_df(transaction_list, return_EUR_values_only = TRUE)
}
\arguments{
\item{transaction_list}{The list containing all transactions from function
get_transactions}

\item{return_EUR_values_only}{When set to TRUE, only EURO-values are considered.
Transactions in foreign currencies are usually converted to EURO-values in a separate
transaction, so no information is lost.}
}
\value{
Dataframe with transactions.
}
\description{
This function transforms the transaction list to a dataframe.
}
